+++ crates/env/src/engine/off_chain/tests.rs
@@ -13,8 +13,13 @@
 // limitations under the License.
 
 use crate::{
-    engine::off_chain::impls::TopicsBuilder,
+    engine::off_chain::{
+        impls::TopicsBuilder,
+        test_api::set_account_balance,
+    },
     event::TopicsBuilderBackend,
+    types::Environment,
+    DefaultEnvironment,
     Result,
 };
 
@@ -37,6 +42,35 @@ fn topics_builder() -> Result<()> {
         #[rustfmt::skip]
         let expected = vec![topics_len_encoded[0], 13, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 17, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0];
         assert_eq!(output, expected);
+
+        Ok(())
+    })
+}
+#[test]
+fn test_set_account_balance() -> Result<()> {
+    pub use ink_engine::ext::ChainSpec;
+
+    crate::test::run_test::<DefaultEnvironment, _>(|_| {
+        let minimum_balance = ChainSpec::default().minimum_balance;
+
+        let result = std::panic::catch_unwind(|| {
+            set_account_balance::<DefaultEnvironment>(
+                <DefaultEnvironment as Environment>::AccountId::from([0x1; 32]),
+                <DefaultEnvironment as Environment>::Balance::from(minimum_balance - 1),
+            )
+        });
+
+        assert!(result.is_err());
+
+        set_account_balance::<DefaultEnvironment>(
+            <DefaultEnvironment as Environment>::AccountId::from([0x1; 32]),
+            <DefaultEnvironment as Environment>::Balance::from(0u128),
+        );
+
+        set_account_balance::<DefaultEnvironment>(
+            <DefaultEnvironment as Environment>::AccountId::from([0x1; 32]),
+            <DefaultEnvironment as Environment>::Balance::from(minimum_balance + 1),
+        );
 
         Ok(())
     })